pdelay/pclock: add pclock state to the standard protocol
pclock is a clock involved in the calculation
of the link delay using Peer Delay Mechanism.
This calculation is independent of whether a clock
is a master/slave. There is delay requester, pclock_req,
and, and delay responder, pclock_resp.
The Grandmaster clock is the only one in charge of sending Sync messages
and complete the synchronization. In case of BC, one
port e.g. port 1 is a pclock connected to the upstream
Grandmaster Clock, and the rest of ports, downstream, Grandmaster
Clocks. In a TC, all the ports are just pclocks, who calculate
the link delay and forward the sync and follow up msg till the slaves.

Signed-off-by: C.Prados <c.prados@gsi.de>