bugfix (provided bmc_dataset_cmp() works fine): distinguis between master port on GM and on BC
master port on GM should copy DefaultDS.clockQuality (e.g. clockClass)
to ParentDS.grandmaster.clockQuality so that it is send out in Announce
messages. BC should not do that, as it copies clockQuality received
on the slave port into the ParentDS.grandmaster.ClockQuality and sends
out on its master ports this info.