p2p:tc: forwarding ann/sync/followup to all up links on the switch
When there is an ingress ann/sync/followup message it is now forwarded
to all enabled ports, acting as a transparent switch to all connected
slaves or TCs.
Copying RX to TX is also now part of a function in common-fun.c for the
three types of message.
Hardcoded values for forwarding from one port to another removed since
now it forwards to all ports now.