p2p: fixing forwarding correction field
cField was not calculated correctly since add_TimeInternal was not
taking  field into account phase field when calculating sync residence
time. On the other hand, all conversions are now correctly done
(pico<->nano).
Also some ingress timestamps and link_delays where overriden wrongly,
now it is fixed.
NOTE: cField is not fully following the standard but since nothing else is using the field, it's fine. This should be fixed in a non-very-far future.