HSR: Transparent Clock behaviour
Adding TC behaviour. WRS can be configured as a transparent clock in
ppsi.conf ("transpclock [enabled|disabled]"). In case it is enabled, the
WRS sintonizes but it never applies PTP to the servo.
Missing: it should be done with clock-class and not in new line. And also variables from the standard probably.