msg.c: temporarily force on a few flags
Our frames shouls announce a ptp timescale, and the fact that both
time and frequency are traceable.

That's true in WR installations but not true in general. Unfortunately
the code base we inherited didn't actually manage propagation of the flags,
and although bmc.c::s1() sets the internal time properties according to
the master we enslave to, such information is never used.

This is a temporary hack, and we really must audit all data sets and frames,
beginning with a rewrite of msg.c based on tools/decent_types.h instead
of hardwired byte offsets.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>