types: replace Boolean with bool C standard type
ieee1588 types as defined in ieee1588_types.h look ambiguous, since
they have the same names of the types defined by the standard, but they
actually refer to internal representation of quantities. For instance
Nibble is defined as uint8_t, so it is ok for representing a nibble, but
has nothing to do with the __actual__ "on-the-wire" nibble.
This patch starts by replacing Boolean with the standard bool C type.