types: introduce struct msg_header_wire and its accessors
We're going to kill MsgHeader and replace it with
struct msg_header_wire, which is a structure representing a message
header in its on-the-wire form (big endian representation,
packed struct, ...).
Accessors are also introduce to read and write message fields in a safe
fashion (we don't have to worry about endianness, bitfields and so on).