std-servo: fix initial jump for p2p slaves
This was my fault. The initial time jump was based on "T4 + MPD". This
on the assumption that T4 is sharper then time->get().  At least this
is true for White Rabbit systems, where time->get() was not even
correct when I started the project.

As a result, the initial jump for a p2p slave was always wrong,
because it referred to an old timestamp. We ended up a few hundreds ms
off the master, to the start a long steering.

We now add ofm to the current time, and the result is good in both case.
WR systems have their own servo anyways, but even a slave to non-wr
masters should behave well, because time->set now works.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>