[HA] added HA-specific calculation of delay_ms and offset
Calculation of delay_ms and offset in HA is different than in in WR
in two ways:
- fixed_delays do not need to be corrected in the delay_ms calculations
- alpha is defined diferently than the fix_alpha

The WR-specific and HA-specific calculations are now in separate
functions, both calculations are run, and the results are printed
while only one of them is used for synchronisation. This is a temporary
solution that helps in development.