[HA] separate function in wr_operations to update all HA correction values.
This function is partially dummy, in the sense that it needs extension
of HAL to provide it with proper data. Now, some values are derived from
the original WR parameters, some are hardcoded.

This commit also cleans up update functions and removes stupid conversion
between delayCoeff and alpha, me bad.