[HA] Use WR servo again, leave HA servo just in case.
It seems that WR servo can be used directly in HA. This is because
different values are defined for WR and HA corrections.
1. when we use WR, deltas are used and ing/egr latencys are zero-ed
2. when we use HA, deltas are zerod and ing/egr latencies are used.
Thus the same core calculations can be used. This fact can be useful
in the future, when we want to have both, HA and WR, working
interchangeably, depending on negotiation.

However, the HA servo has some debugging in it and can be useful in
the future, so leave it but not compile.

Using back WR servo required removal of some functions that I moved
out of servo.