[HA] replaced the variables and bitmasks defined in wr_dsport by data sets and their members defined in the standard.
The standard defines data sets and members to store bits/values for
L1Sync. These should not be part of wr_dsport inside PortDS. They are
at the same level as PortDS. Also. I did not like the bitmasks stored
and the macros used later to set/get/read them. This commit makes the
implementation follow directly the standard. It also adds some functions
to create bitmask out of DS members values and to print the info about
active and configuration bits in human-readible way (not hex).